requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.23.2
jinja2>=3.1.2
python-multipart>=0.0.6

# Data Handling
pydantic>=2.4.2
pydantic[email]>=2.4.2
python-dotenv>=1.0.0
sqlalchemy>=2.0.22

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
pyotp>=2.9.0

# Error Handling & Retries
tenacity>=8.2.2

# API Integrations
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.3
google-auth-oauthlib>=1.1.0
todoist-api-python>=2.1.3

# Task Scheduling
apscheduler>=3.10.4

# LLM Integration
anthropic>=0.5.0  # For Deepseek R1 or equivalent

# Testing
pytest>=7.4.2
pytest-cov>=4.1.0

# Linting and Type Checking
flake8>=6.1.0
black>=23.9.1
mypy>=1.6.1

# Docker Support
docker>=6.1.3